+++ image_transport/include/image_transport/subscriber_filter.hpp
@@ -138,7 +138,7 @@ public:
    * \brief Returns the number of publishers this subscriber is connected to.
    */
   IMAGE_TRANSPORT_PUBLIC
-  uint32_t getNumPublishers() const
+  size_t getNumPublishers() const
   {
     return sub_.getNumPublishers();
   }
